requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-MySQL==1.5.2
Flask-WTF==0.14.3
nltk==3.5
numpy==1.19.0
pandas==1.1.5
plotly==4.5.0
python-dotenv==0.17.0
requests==2.25.1
Sastrawi==1.0.1
wordcloud==1.8.1
xlrd==1.2.0
